+++ src/common.h
@@ -48,4 +48,4 @@
 int is_valid_ip(char *ip_string);
 int is_valid_port(char *port_string);
 int starts_with(const char *a, const char *b);
-int min(int a, int b);
+size_t min(size_t a, size_t b);
